Move api/server/router/system to daemon/server/router/system
