+++ daemon/server/router/system/info_response.go
